src/requirements.txt
elasticsearch==8.13.0
google-cloud-secret-manager==2.22.0
